+++ taf/updater/updater_pipeline.py
@@ -2,15 +2,15 @@ from collections import defaultdict
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from enum import Enum
 import functools
-from logging import DEBUG, INFO
 from pathlib import Path
 import re
 import shutil
 import tempfile
 from typing import Any, Dict, List, Optional
+
 from attr import attrs, define, field
+
 from taf.git import GitError
-from logdecorator import log_on_end, log_on_start
 from taf.git import GitRepository
 from taf.constants import INFO_JSON_PATH
 
@@ -28,9 +28,8 @@ from taf.updater.handlers import GitUpdater
 from taf.updater.lifecycle_handlers import Event
 from taf.updater.types.update import OperationType, UpdateType
 from taf.utils import TempPartition, on_rm_error, ensure_pre_push_hook
-from taf.log import taf_logger
 from tuf.ngclient.updater import Updater
-
+from taf.log import taf_logger
 
 EXPIRED_METADATA_ERROR = "ExpiredMetadataError"
 
@@ -197,6 +196,11 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         super().__init__(
             steps=[
                 (
+                    self.start_update,
+                    RunMode.ALL,
+                    None,
+                ),
+                (
                     self.set_existing_repositories,
                     RunMode.UPDATE,
                     None,
@@ -289,6 +293,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     self.should_validate_target_repos,
                 ),  # skipped with no-targets; prints all other commits that exist but are not merged
                 (self.check_pre_push_hook, RunMode.ALL, self.should_update_auth_repos),
+                (self.finish_update, RunMode.ALL, None),
             ],
             run_mode=(
                 RunMode.LOCAL_VALIDATION
@@ -337,10 +342,26 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             return self.state.event == Event.CHANGED
         return True
 
-    @log_on_start(
-        DEBUG, "Checking which repositories are already on disk...", logger=taf_logger
-    )
+    def start_update(self):
+        # This message should be shown regardless of verbosity setting
+        update_text = "validation" if self.only_validate else "update"
+        if self.auth_path:
+            auth_repo_name = GitRepository(path=self.auth_path).name
+            taf_logger.log("NOTICE", f"{auth_repo_name}: Starting {update_text}...")
+        else:
+            taf_logger.log("NOTICE", f"Starting {update_text}...")
+
+    def finish_update(self):
+        # This message should be shown regardless of verbosity setting
+        update_text = "validation" if self.only_validate else "update"
+        taf_logger.log(
+            "NOTICE", f"{self.state.auth_repo_name}: Finished {update_text}!"
+        )
+
     def set_existing_repositories(self):
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Checking which repositories are already on disk..."
+        )
         self.state.existing_repo = False
         self.state.repos_on_disk = {}
         if self.auth_path is not None:
@@ -368,16 +389,16 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         return UpdateStatus.SUCCESS
 
     # return UpdateStatus.SUCCESS if self.state.existing_repo else UpdateStatus.FAILURE
-    @log_on_start(
-        INFO,
-        "Checking if local repositories are clean...",
-        logger=taf_logger,
-    )
     def check_if_local_repositories_clean(self):
         try:
             # check if the auth repo is clean first
             if self.state.existing_repo:
-                auth_repo = AuthenticationRepository(path=self.auth_path)
+                auth_repo = AuthenticationRepository(
+                    path=self.auth_path, urls=[self.url]
+                )
+                taf_logger.info(
+                    f"{auth_repo.name}: Checking if local repositories are clean..."
+                )
                 if auth_repo.is_bare_repository:
                     taf_logger.info(
                         f"Skipping clean check for bare repository {auth_repo.name}"
@@ -451,11 +472,15 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(
-        INFO, "Cloning repository and running TUF updater...", logger=taf_logger
-    )
     @cleanup_decorator
     def clone_remote_and_run_tuf_updater(self):
+        if self.auth_path:
+            auth_repo_name = GitRepository(path=self.auth_path).name
+            taf_logger.info(
+                f"{auth_repo_name}: Cloning repository and running TUF updater..."
+            )
+        else:
+            taf_logger.info("Cloning repository and running TUF updater...")
         settings.update_from_filesystem = self.update_from_filesystem
 
         if self.operation == OperationType.CLONE_OR_UPDATE:
@@ -600,12 +625,12 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 f"{self.state.users_auth_repo.path} is not a Git repository. Run 'taf repo clone' instead"
             )
 
-    @log_on_start(
-        INFO, "Validating out of band commit and update type", logger=taf_logger
-    )
     def validate_out_of_band_and_update_type(self):
         # this is the repository cloned inside the temp directory
         # we validate it before updating the actual authentication repository
+        taf_logger.info(
+            f"{self.state.auth_repo_name}: Validating out of band commit and update type..."
+        )
         try:
             if (
                 self.out_of_band_authentication is not None
@@ -705,14 +730,12 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             self.state.users_auth_repo.set_last_validated_commit(users_head_sha)
             self.state.last_validated_commit = users_head_sha
 
-    @log_on_start(
-        INFO,
-        "Cloning or updating user's authentication repository...",
-        logger=taf_logger,
-    )
     def clone_or_fetch_users_auth_repo(self):
         # fetch the latest commit or clone the repository without checkout
         # do not merge before targets are validated as well
+        taf_logger.info(
+            f"{self.state.auth_repo_name}: Cloning or updating user's authentication repository..."
+        )
         try:
             if self.state.existing_repo:
                 self.state.users_auth_repo.fetch(fetch_all=True)
@@ -727,8 +750,8 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             return UpdateStatus.FAILED
         return UpdateStatus.SUCCESS
 
-    @log_on_start(DEBUG, "Loading target repositories", logger=taf_logger)
     def load_target_repositories(self):
+        taf_logger.debug(f"{self.state.auth_repo_name}: Loading target repositories...")
         try:
             repositoriesdb.load_repositories(
                 self.state.users_auth_repo,
@@ -766,12 +789,10 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(
-        INFO,
-        "Checking if all target repositories are already on disk...",
-        logger=taf_logger,
-    )
     def check_if_repositories_on_disk(self):
+        taf_logger.info(
+            f"{self.state.auth_repo_name}: Checking if all target repositories are already on disk..."
+        )
         try:
             for repository in self.state.users_target_repositories.values():
                 if not repository.is_git_repository_root:
@@ -788,9 +809,10 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(DEBUG, "Cloning target repositories to temp...", logger=taf_logger)
-    @log_on_end(INFO, "Finished cloning target repositories", logger=taf_logger)
     def clone_target_repositories_to_temp(self):
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Cloning target repositories to temp..."
+        )
         try:
             self.state.repos_on_disk = {}
             self.state.repos_not_on_disk = {}
@@ -817,22 +839,19 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
                 for future in as_completed(futures):
                     future.result()
-
+            taf_logger.info(
+                f"{self.state.auth_repo_name}: Finished cloning target repositories."
+            )
             return UpdateStatus.SUCCESS
         except Exception as e:
             self.state.errors.append(e)
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(
-        INFO, "Validating initial state of target repositories...", logger=taf_logger
-    )
-    @log_on_end(
-        INFO,
-        "Checking initial state of repositories",
-        logger=taf_logger,
-    )
     def determine_start_commits(self):
+        taf_logger.info(
+            f"{self.state.auth_repo_name}: Validating initial state of target repositories..."
+        )
         try:
             self.state.targets_data_by_auth_commits = (
                 self.state.users_auth_repo.targets_data_by_auth_commits(
@@ -871,7 +890,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
             if not is_initial_state_in_sync:
                 taf_logger.info(
-                    f"Repository {self.state.users_auth_repo.name}: states of target repositories are not in sync with last validated commit. Starting the update from the beginning"
+                    f"{self.state.users_auth_repo.name}: states of target repositories are not in sync with last validated commit. Starting the update from the beginning"
                 )
                 self._update_state_for_initial_sync()
                 self.reset_target_repositories()
@@ -956,9 +975,11 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         self.state.target_branches_data_from_auth_repo = repo_branches
         return UpdateStatus.SUCCESS
 
-    @log_on_start(DEBUG, "Fetching commits of target repositories", logger=taf_logger)
     def get_target_repositories_commits(self):
         """Returns a list of newly fetched commits belonging to the specified branch."""
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Fetching commits of target repositories..."
+        )
         self.state.fetched_commits_per_target_repos_branches = defaultdict(dict)
 
         def fetch_commits(repository, branch, old_head):
@@ -1048,12 +1069,10 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     return UpdateStatus.FAILED
         return UpdateStatus.SUCCESS
 
-    @log_on_start(
-        DEBUG,
-        "Checking if target repositories are clean...",
-        logger=taf_logger,
-    )
     def check_if_local_target_repositories_clean(self):
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Checking if target repositories are clean..."
+        )
         try:
             for repository in self.state.repos_on_disk.values():
                 if repository.is_bare_repository:
@@ -1080,13 +1099,14 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             return UpdateStatus.FAILED
         return UpdateStatus.SUCCESS
 
-    @log_on_start(INFO, "Validating target repositories...", logger=taf_logger)
-    @log_on_end(INFO, "Validation of target repositories finished", logger=taf_logger)
     def validate_target_repositories(self):
         """
         Breadth-first update of target repositories
         Merge last valid commits at the end of the update
         """
+        taf_logger.info(
+            "f{self.state.auth_repo_name}: Validating target repositories..."
+        )
         try:
             # need to be set to old head since that is the last validated target
             self.state.validated_commits_per_target_repos_branches = defaultdict(dict)
@@ -1155,6 +1175,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
                 # commit processed without an error
                 self.state.validated_auth_commits.append(auth_commit)
+            taf_logger.info(
+                f"{self.state.auth_repo_name}: Validation of target repositories finished"
+            )
             return UpdateStatus.SUCCESS
         except Exception as e:
             self.state.errors.append(e)
@@ -1228,11 +1251,6 @@ data repository {repository.name} was supposed to be at commit {current_commit} 
 but commit not on branch {current_branch}"
         )
 
-    @log_on_start(
-        DEBUG,
-        "Validating and setting additional commits of target repositories",
-        logger=taf_logger,
-    )
     def validate_and_set_additional_commits_of_target_repositories(self):
         """
         For target repository and for each branch, extract commits following the last validated commit
@@ -1241,6 +1259,9 @@ but commit not on branch {current_branch}"
         However, no error will get reported if there are commits which have not yet been signed
         In case of repositories which can contain unauthenticated commits, they do not even need to get signed
         """
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Validating and setting additional commits of target repositories..."
+        )
         # only get additional commits if the validation was complete (not partial, up to a commit)
         self.state.additional_commits_per_target_repos_branches = defaultdict(dict)
         if self.state.update_status != UpdateStatus.SUCCESS:
@@ -1296,10 +1317,10 @@ but commit not on branch {current_branch}"
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(
-        DEBUG, "Copying or updating user's target repositories...", logger=taf_logger
-    )
     def update_users_target_repositories(self):
+        taf_logger.debug(
+            f"{self.state.auth_repo_name}: Copying or updating user's target repositories..."
+        )
         if self.state.update_status == UpdateStatus.FAILED:
             return self.state.update_status
         try:
@@ -1330,8 +1351,8 @@ but commit not on branch {current_branch}"
             self.state.event = Event.FAILED
             return UpdateStatus.FAILED
 
-    @log_on_start(DEBUG, "Removing temp repositories...", logger=taf_logger)
     def remove_temp_repositories(self):
+        taf_logger.debug(f"{self.state.auth_repo_name}: Removing temp repositories...")
         if not self.state.temp_root:
             return self.state.update_status
         try:
@@ -1344,13 +1365,13 @@ but commit not on branch {current_branch}"
             )
         return self.state.update_status
 
-    @log_on_start(
-        INFO, "Merging commits into target repositories...", logger=taf_logger
-    )
     def merge_commits(self):
         """Determines which commits needs to be merged into the specified branch and
         merge it.
         """
+        taf_logger.info(
+            f"{self.state.auth_repo_name}: Merging commits into target repositories..."
+        )
         try:
             if self.only_validate:
                 return self.state.update_status
@@ -1367,7 +1388,7 @@ but commit not on branch {current_branch}"
                     last_validated_commit = validated_commits[-1]
                     commit_to_merge = last_validated_commit
                     _merge_commit(
-                        repository, branch, commit_to_merge, force_revert=self.force
+                        repository, branch, commit_to_merge, force_revert=True
                     )
             return self.state.update_status
         except Exception as e:
@@ -1535,12 +1556,10 @@ def _is_unauthenticated_allowed(repository):
     return repository.custom.get("allow-unauthenticated-commits", False)
 
 
-@log_on_start(
-    INFO,
-    "Running TUF validation of the authentication repository...",
-    logger=taf_logger,
-)
 def _run_tuf_updater(git_fetcher, auth_repo_name):
+    auth_repo_name = auth_repo_name or ""
+    taf_logger.info(f"{auth_repo_name}: Running TUF validation...")
+
     def _init_updater():
         try:
             return Updater(
@@ -1551,7 +1570,9 @@ def _run_tuf_updater(git_fetcher, auth_repo_name):
                 fetcher=git_fetcher,
             )
         except Exception as e:
-            taf_logger.error(f"Failed to instantiate TUF Updater due to error: {e}")
+            taf_logger.error(
+                f"{auth_repo_name}: Failed to instantiate TUF Updater due to error: {e}"
+            )
             raise e
 
     last_validated_commit = None
@@ -1572,8 +1593,11 @@ def _run_tuf_updater(git_fetcher, auth_repo_name):
 def _update_tuf_current_revision(git_fetcher, updater, auth_repo_name):
     current_commit = git_fetcher.current_commit
     try:
+        auth_repo_name = auth_repo_name or ""
         updater.refresh()
-        taf_logger.debug("Validated metadata files at revision {}", current_commit)
+        taf_logger.debug(
+            f"{auth_repo_name}: Validated metadata files at revision {current_commit}"
+        )
         # using refresh, we have updated all main roles
         # we still need to update the delegated roles (if there are any)
         # and validate any target files
@@ -1585,9 +1609,7 @@ def _update_tuf_current_revision(git_fetcher, updater, auth_repo_name):
             targetinfo.verify_length_and_hashes(target_data)
 
             taf_logger.debug(
-                "Successfully validated target file {} at {}",
-                target_filepath,
-                current_commit,
+                f"{auth_repo_name}: Successfully validated target file {target_filepath} at {current_commit}"
             )
         if settings.strict:
             _validate_metadata_on_disk(git_fetcher)
@@ -1598,17 +1620,14 @@ def _update_tuf_current_revision(git_fetcher, updater, auth_repo_name):
         ).__name__ or EXPIRED_METADATA_ERROR in str(e)
         if not metadata_expired or settings.strict:
             taf_logger.error(
-                "Validation of authentication repository {} failed at revision {} due to error: {}",
-                auth_repo_name or "",
-                current_commit,
-                e,
+                f"{auth_repo_name}: Validation of authentication repository failed at revision {current_commit} due to error: {e}"
             )
             raise UpdateFailedError(
                 f"Validation of authentication repository {auth_repo_name or ''}"
                 f" failed at revision {current_commit} due to error: {e}"
             )
         taf_logger.warning(
-            f"WARNING: Could not validate authentication repository at revision {current_commit} due to error: {e}"
+            f"WARNING: Could not validate authentication repository {auth_repo_name} at revision {current_commit} due to error: {e}"
         )